+++ source/sites/publicodes/SubCategoriesChart.tsx
@@ -1,32 +1,11 @@
 import { AnimatePresence, motion } from 'framer-motion'
 import emoji from 'react-easy-emoji'
-import { useSelector } from 'react-redux'
 import styled from 'styled-components'
-import { useEngine } from '../../components/utils/EngineContext'
 import { extractCategories } from 'Components/publicodesUtils'
 import SubCategoryBar from './SubCategoryBar'
-import { ruleFormula } from '../../components/publicodesUtils'
 import { shadowStyle } from './styles'
 
-export default ({ category, color }) => {
-	const rules = useSelector((state) => state.rules)
-	const engine = useEngine()
-	const evaluated = engine.evaluate(category),
-		total = evaluated.nodeValue,
-		rule = engine.getRule(category),
-		formula = ruleFormula(rule)
-
-	if (!formula) return null
-
-	const sumToDisplay =
-		formula.nodeKind === 'somme'
-			? category
-			: formula.operationKind === '/'
-			? formula.explanation[0].dottedName
-			: null
-
-	if (!sumToDisplay) return null
-
+export default ({ rules, engine, sumToDisplay, color, total }) => {
 	const subCategories = extractCategories(
 		rules,
 		engine,
@@ -44,7 +23,6 @@ export default ({ category, color }) => {
 
 	return (
 		<div>
-			{formula.operationKind === '/' && <span>Salut</span>}
 			<InlineBarChart>
 				<AnimatePresence>
 					{subCategories.map(({ nodeValue, title, icons }) => (
